rename data_parser subpackage to io
