+++ artist/__init__.py
@@ -2,14 +2,14 @@
 
 import os
 
-from . import core, data_parser, field, scenario, scene, util
+from . import core, field, io, scenario, scene, util
 
 ARTIST_ROOT = f"{os.sep}".join(__file__.split(os.sep)[:-2])
 """Reference to the root directory of ``ARTIST``."""
 
 __all__ = [
     "core",
-    "data_parser",
+    "io",
     "field",
     "scenario",
     "scene",
