+++ artist/core/surface_reconstructor.py
@@ -9,8 +9,8 @@ from artist.core import core_utils, learning_rate_schedulers
 from artist.core.heliostat_ray_tracer import HeliostatRayTracer
 from artist.core.loss_functions import Loss
 from artist.core.regularizers import IdealSurfaceRegularizer, SmoothnessRegularizer
-from artist.data_parser.calibration_data_parser import CalibrationDataParser
 from artist.field.heliostat_group import HeliostatGroup
+from artist.io.calibration_data_parser import CalibrationDataParser
 from artist.scenario.scenario import Scenario
 from artist.util import (
     config_dictionary,
