+++ artist/io/h5_scenario_parser.py
