+++ artist/io/paint_scenario_parser.py
