+++ artist/io/stral_scenario_parser.py
