+++ examples/paint_plots/reconstruction_generate_results.py
@@ -12,8 +12,8 @@ import yaml
 
 from artist.core.kinematics_reconstructor import KinematicsReconstructor
 from artist.core.loss_functions import FocalSpotLoss
-from artist.data_parser.calibration_data_parser import CalibrationDataParser
-from artist.data_parser.paint_calibration_parser import PaintCalibrationDataParser
+from artist.io.calibration_data_parser import CalibrationDataParser
+from artist.io.paint_calibration_parser import PaintCalibrationDataParser
 from artist.scenario import Scenario
 from artist.util import config_dictionary, set_logger_config
 from artist.util.environment_setup import (
