+++ tests/scenario/test_surface_generator.py
@@ -4,7 +4,7 @@ import pytest
 import torch
 
 from artist import ARTIST_ROOT
-from artist.data_parser import paint_scenario_parser, stral_scenario_parser
+from artist.io import paint_scenario_parser, stral_scenario_parser
 from artist.scenario.configuration_classes import FacetConfig, SurfaceConfig
 from artist.scenario.surface_generator import SurfaceGenerator
 from artist.util import config_dictionary
