+++ tutorials/03_nurbs_surface_reconstruction_tutorial.py
@@ -10,9 +10,9 @@ from matplotlib import pyplot as plt
 from artist.core.heliostat_ray_tracer import HeliostatRayTracer
 from artist.core.loss_functions import KLDivergenceLoss
 from artist.core.surface_reconstructor import SurfaceReconstructor
-from artist.data_parser import paint_scenario_parser
-from artist.data_parser.calibration_data_parser import CalibrationDataParser
-from artist.data_parser.paint_calibration_parser import PaintCalibrationDataParser
+from artist.io import paint_scenario_parser
+from artist.io.calibration_data_parser import CalibrationDataParser
+from artist.io.paint_calibration_parser import PaintCalibrationDataParser
 from artist.scenario.scenario import Scenario
 from artist.util import config_dictionary, index_mapping, set_logger_config, utils
 from artist.util.environment_setup import get_device, setup_distributed_environment
